backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
firebase-admin==6.2.0
pymongo==4.6.0
motor==3.3.2
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4
nltk==3.8.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
cors-middleware==1.0.1
sendgrid==6.10.0
celery==5.3.4
redis==5.0.1
tenacity==8.2.3
requests==2.31.0
boto3==1.29.7
google-cloud-storage==2.14.0
google-cloud-logging==3.8.0
python-dateutil==2.8.2
